pyproject.toml
[project]
name = "domain-intelligence"
version = "1.0.0"
description = "External asset discovery and threat assessment tool for organizational domains"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Security Team"}
]
keywords = ["security", "domain", "reconnaissance", "threat-intelligence", "dns", "ssl"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: System :: Networking",
]

dependencies = [
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "dnspython>=2.3.0",
    "python-whois>=0.8.0",
    "cryptography>=40.0.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "types-requests",
    "types-PyYAML",
]

[project.scripts]
domain-intel = "cli:main"

[project.urls]
Homepage = "https://github.com/your-org/domain-intelligence"
Documentation = "https://github.com/your-org/domain-intelligence#readme"
Issues = "https://github.com/your-org/domain-intelligence/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
